+++ src/panelWebView/Command.ts
@@ -10,6 +10,5 @@ export enum Command {
   sendMediaUrl = 'sendMediaUrl',
   updatePlaceholder = 'updatePlaceholder',
   dataFileEntries = 'dataFileEntries',
-  updatedSlug = 'updatedSlug',
   serverStarted = 'server-started'
 }